backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
pdf2docx==0.5.8
PyMuPDF==1.24.9
httpx==0.27.2
groq==0.11.0
aiofiles==23.2.1
python-dotenv==1.0.1
